Moved Image.jsx and PhotoLink.jsx in a folder along with respective css file. Added css to image and image buttons
