+++ photo-effects/src/components/Dashboard/DashNav/dashNav.css
@@ -1,12 +1,10 @@
+
 .dashNav {
-    border-bottom: 1px solid #ffdb27;
-    padding: 10px;
     display: flex;
     justify-content: space-between;
     align-items: center;
     width: 100%;
-    height: 11vh;
-    padding: 10px;
+    height: 14vh;
     background-color: #364f6b;
     color: whitesmoke;
     font-size: 2rem;
@@ -31,7 +29,6 @@
   }
 
   .home {
-      border: 1px solid black;
       cursor: pointer;
       padding: 5px;
   }
@@ -56,4 +53,20 @@
       filter: brightness(110%);
       transform: scale(1.1);
       transition: ease-in-out 0.1s
+  }
+
+  /* Dashboard */
+  .welcome {
+      padding-top: 16vh;
+      font-size: 3.5rem;
+  }
+
+  .center {
+      text-align: center;
+      margin-top: 50px;
+      /* border: 1px solid black; */
+  }
+
+  .center h2 {
+    font-size: 4.6rem;
   }
