+++ photo-effects/src/components/Dashboard/Image/Image.jsx
