+++ photo-effects/src/components/Dashboard/Image/image.css
@@ -1,0 +1,35 @@
+
+
+.image-link {
+    margin-top: 10px;
+}
+
+.btn-center {
+   display: flex;
+   justify-content: center;
+   margin-top: 10px;
+
+}
+
+.btn-flex {
+    border: 1px solid black;
+    display: flex;
+    justify-content: space-between;
+    width: 40%;
+}
+
+.btn {
+    background-color: #fc5185;
+    padding: 12px;
+    border-radius: 8px;
+    color: whitesmoke;
+    box-shadow: -1px 1px 90px -2px rgba(240,240,240,1);
+    border: none;
+    font-size:2rem;
+    cursor: pointer;
+}
+.btn:hover {
+    filter: brightness(110%);
+    transform: scale(1.1);
+    transition: ease-in-out 0.1s
+}
