--- photo-effects/src/components/Dashboard/image.css
@@ -1,6 +1,0 @@
-
-
-
-.remove {
-    font-size: 100px;
-}
